Refactor - Native Classes blueprints: rename to `edition-detector`
