+++ blueprints/component/index.js
@@ -5,11 +5,11 @@ const stringUtil = require('ember-cli-string-utils');
 const pathUtil = require('ember-cli-path-utils');
 const getPathOption = require('ember-cli-get-component-path-option');
 const normalizeEntityName = require('ember-cli-normalize-entity-name');
-const useNativeDetector = require('../native-detector');
+const useEditionDetector = require('../edition-detector');
 const isModuleUnificationProject = require('../module-unification').isModuleUnificationProject;
 const EOL = require('os').EOL;
 
-module.exports = useNativeDetector({
+module.exports = useEditionDetector({
   description: 'Generates a component.',
 
   availableOptions: [
