+++ blueprints/edition-detector.js
