+++ blueprints/route/index.js
@@ -5,10 +5,10 @@ const path = require('path');
 const chalk = require('chalk');
 const stringUtil = require('ember-cli-string-utils');
 const EmberRouterGenerator = require('ember-router-generator');
-const useNativeDetector = require('../native-detector');
+const useEditionDetector = require('../edition-detector');
 const isModuleUnificationProject = require('../module-unification').isModuleUnificationProject;
 
-module.exports = useNativeDetector({
+module.exports = useEditionDetector({
   description: 'Generates a route and a template, and registers the route with the router.',
 
   availableOptions: [
